remove SRPSession/SRPClientSession inheritance indirection
doesn't add any value, inheritance was only used to split code
